+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Unreleased
     * Add `HelloTm1636`, `HelloMax7219`, `HelloHt16k33`, and `HelloHc595`
       examples to `examples/` and `README.md`.
+    * Clean up methods to retrieve underlying `patternWriter()`, `charWriter()`,
+      `numberWriter()`, etc in various Writer classes.
 * 0.8 (2021-07-30)
     * Move I2C address into `beginTransmission()` method of `TwoWireInterface`,
       `SimpleWireInterface` and `SimpleWireFastInterface`.